Jenny_Login_Feature

+++ EmployeeDirectory/Data/Entity/Employee.vb
@@ -3,12 +3,12 @@
 
     'Employee Information
 
-    Private _OracleID As Integer
-    Public Property OracleID() As Integer
+    Private _OracleID As Long
+    Public Property OracleID() As Long
         Get
             Return _OracleID
         End Get
-        Set(ByVal value As Integer)
+        Set(ByVal value As Long)
             _OracleID = value
         End Set
     End Property
@@ -134,12 +134,12 @@
         End Set
     End Property
 
-    Private _LocalManagerID As Integer
-    Public Property LocalManagerID() As Integer
+    Private _LocalManagerID As Long
+    Public Property LocalManagerID() As Long
         Get
             Return _LocalManagerID
         End Get
-        Set(ByVal value As Integer)
+        Set(ByVal value As Long)
             _LocalManagerID = value
         End Set
     End Property
